requirements.txt
numpy==1.21.5
scipy==1.8.0
openslide_python==1.3.1
dask==2025.3.0
matplotlib==3.9.2
dask-image==2024.5.3
